+++ src/modules/revision/revision.controller.ts
@@ -4,14 +4,11 @@ import {
   HttpException,
   HttpStatus,
   ParseArrayPipe,
-  Post,
   Query,
   Req,
   Res,
-  UseGuards,
 } from '@nestjs/common';
 import {
-  ApiBearerAuth,
   ApiOperation,
   ApiParam,
   ApiQuery,
@@ -30,9 +27,6 @@ import { RevisionWithClientDTO } from './dto/revision_with_client.dto';
 import { RevisionQueryDto } from './dto/status_revisions.dto';
 import { AnciennesCommunesDTO } from './dto/ancienne_commune.dto';
 import { ValidationCogPipe } from '@/lib/class/pipes/validation_cog.pipe';
-import { AdminGuard } from '@/lib/class/guards/admin.guard';
-import { ClientService } from '../client/client.service';
-import { ConfigService } from '@nestjs/config';
 
 @ApiTags('revisions')
 @Controller('')
@@ -40,8 +34,6 @@ export class RevisionController {
   constructor(
     private revisionService: RevisionService,
     private fileService: FileService,
-    private clientService: ClientService,
-    private configService: ConfigService,
   ) {}
 
   @Get('current-revisions')
@@ -217,47 +209,5 @@ export class RevisionController {
     res.attachment(`bal-${req.revision.codeCommune}.csv`);
     res.setHeader('Content-Type', 'text/csv');
     res.send(data);
-  }
-
-  @Post('revisions/:revisionId/sync-ids-ban-publish')
-  @ApiParam({ name: 'revisionId', required: true, type: String })
-  @ApiOperation({
-    summary: 'Synchro ids BAL with ids BAN and publish',
-    operationId: 'syncIdsBANPublish',
-  })
-  @ApiBearerAuth('admin-token')
-  @UseGuards(AdminGuard)
-  @ApiResponse({
-    status: HttpStatus.OK,
-    type: Revision,
-  })
-  async syncIdsBANPublish(@Req() req: CustomRequest, @Res() res: Response) {
-    try {
-      const { file: csvFileSync, codeCommune } =
-        await this.revisionService.syncIdsBAN(req.revision);
-
-      const client = await this.clientService.findOneOrFail(
-        req.revision.clientId,
-      );
-      const newRevisionFirstStep: Revision =
-        await this.revisionService.createOne(codeCommune, client, {
-          ...req.revision.context,
-          extras: {
-            ...req.revision.context?.extras,
-            syncRevisionId: req.revision.id,
-          },
-        });
-      await this.revisionService.setFile(newRevisionFirstStep, csvFileSync);
-      const newRevisionSecondStep: Revision =
-        await this.revisionService.computeOne(newRevisionFirstStep, client);
-      const newRevisionFinalStep: Revision =
-        await this.revisionService.publishOneWithLock(
-          newRevisionSecondStep,
-          client,
-        );
-      res.status(HttpStatus.OK).send(newRevisionFinalStep);
-    } catch (error) {
-      throw new HttpException(error.message, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
   }
 }
